refactor: use barrel file for inputs
